Fixed the import-cycle Refactored Incidents creation to a classmetod, added the update_incident as static_method to incidents Removed the updates.py
