--- app/incident/__init__.py
@@ -1,3 +1,0 @@
-from app.incident.incident import Incident
-from app.incident.incidents import Incidents
-from app.incident.updates import update_incident, actual_version
